+++ myqueue/local.py
@@ -2,7 +2,7 @@ import asyncio
 import pickle
 import socket
 from functools import partial
-from typing import Any, Set, Tuple, List
+from typing import Any, Set, Tuple, List, Dict
 
 from .scheduler import Scheduler
 from .task import Task
@@ -15,6 +15,8 @@ class LocalSchedulerError(Exception):
 
 
 class LocalScheduler(Scheduler):
+    port = 39999
+
     def submit(self,
                task: Task,
                dry_run: bool = False,
@@ -23,7 +25,7 @@ class LocalScheduler(Scheduler):
             id = 1
         else:
             task.cmd.function = None
-            (id,) = self.send('submit', task)
+            id = self.send('submit', task)
         task.id = id
 
     def cancel(self, task: Task) -> None:
@@ -36,13 +38,13 @@ class LocalScheduler(Scheduler):
         self.send('release', task.id)
 
     def get_ids(self) -> Set[int]:
-        (ids,) = self.send('list')
+        ids = self.send('list')
         return ids
 
-    def send(self, *args: Any) -> Tuple[Any, ...]:
+    def send(self, *args: Any) -> Any:
         with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
             try:
-                s.connect(('127.0.0.1', 8888))
+                s.connect(('127.0.0.1', self.port))
             except ConnectionRefusedError:
                 raise ConnectionRefusedError(
                     'Local scheduler not responding.  '
@@ -50,11 +52,13 @@ class LocalScheduler(Scheduler):
             b = pickle.dumps(args)
             assert len(b) < 4096
             s.sendall(b)
+            if args[0] == 'stop':
+                return
             b = b''.join(iter(partial(s.recv, 4096), b''))
-        status, *args = pickle.loads(b)
+        status, result = pickle.loads(b)
         if status != 'ok':
             raise LocalSchedulerError(status)
-        return args
+        return result
 
     def get_config(self, queue: str = '') -> Tuple[List[Tuple[str, int, str]],
                                                    List[str]]:
@@ -62,52 +66,94 @@ class LocalScheduler(Scheduler):
 
 
 class Server:
-    def __init__(self) -> None:
+    def __init__(self,
+                 config: Configuration,
+                 cores: int = 1,
+                 port: int = 39999) -> None:
+        self.config = config
+        self.port = port
+
         self.next_id = 1
-        self.tasks: List[Task] = []
-        self.config = Configuration.read()
+        self.tasks: Dict[int, Task] = {}
+        self.processes: Dict[int, asyncio.subprocess.Process] = {}
+        self.aiotasks: Dict[int, asyncio.Task] = {}
         self.folder = self.config.home / '.myqueue'
 
     async def main(self) -> None:
-        server = await asyncio.start_server(
-            self.recv, '127.0.0.1', 8888)
+        self.server = await asyncio.start_server(
+            self.recv, '127.0.0.1', self.port)
 
-        # self.task = asyncio.create_task(self.execute())
+        async with self.server:  # type: ignore
+            await self.server.serve_forever()
+            await self.server.wait_closed()
 
-        async with server:  # type: ignore
-            await server.serve_forever()
+    def start(self) -> None:
+        try:
+            asyncio.run(self.main())
+        except asyncio.exceptions.CancelledError:
+            pass
 
     async def recv(self, reader: Any, writer: Any) -> None:
-
         data = await reader.read(4096)
         cmd, *args = pickle.loads(data)
-        print(cmd, args)
-        if cmd == 'submit':
+        print('COMMAND:', cmd, *args)
+        result: Any
+        if cmd == 'stop':
+            for aiotask in self.aiotasks.values():
+                await aiotask
+            self.server.close()
+            print('BYE')
+            return
+            result = None
+        elif cmd == 'submit':
             task = args[0]
             task.id = self.next_id
+            task.state = State.queued
+            task.deps = [t.dname for t in task.dtasks
+                         if t.id in self.tasks]
             self.next_id += 1
-            self.tasks.append(task)
-            result = (task.id,)
+            self.tasks[task.id] = task
+            result = task.id
+        elif cmd == 'list':
+            result = list(self.tasks)
+        elif cmd == 'cancel':
+            id = args[0]
+            if id in self.processes:
+                self.terminate(id)
+            elif id in self.tasks:
+                task = self.tasks[id]
+                task.state = State.CANCELED
+                task.cancel_dependents(list(self.tasks.values()))
+                self.tasks = {id: task for id, task in self.tasks.items()
+                              if task.state != State.CANCELED}
+            result = None
         else:
             1 / 0
-        writer.write(pickle.dumps(('ok',) + result))
+        writer.write(pickle.dumps(('ok', result)))
         await writer.drain()
         writer.close()
         self.kick()
-        print(self.next_id, self.tasks)
+        print('JOBS:', list(self.tasks))
 
     def kick(self) -> None:
-        for task in self.tasks:
-            if task.state == 'running':
+        self.aiotasks = {id: aiotask
+                         for id, aiotask in self.aiotasks.items()
+                         if not aiotask.done()}
+
+        for task in self.tasks.values():
+            if task.state == State.running:
                 return
 
-        for task in self.tasks:
-            if task.state == 'queued' and not task.deps:
+        for task in self.tasks.values():
+            if task.state == State.queued and not task.deps:
                 break
-        else:
+        else:  # no break
             return
 
-        asyncio.create_task(self.run(task))
+        print('START', task.id)
+        aiotask = asyncio.create_task(self.run(task))
+        self.aiotasks[task.id] = aiotask
+        # aiotask.add_done_callback(lambda t: self.aiotasks.pop(task.id))
 
     async def run(self, task: Task) -> None:
         out = f'{task.cmd.short_name}.{task.id}.out'
@@ -120,17 +166,27 @@ class Server:
             cmd = mpiexec + cmd.replace('python3',
                                         self.config.parallel_python)
         cmd = f'{cmd} 2> {err} > {out}'
+
         proc = await asyncio.create_subprocess_shell(
             cmd, cwd=task.folder)
+
+        self.processes[task.id] = proc
+
         loop = asyncio.get_event_loop()
         tmax = task.resources.tmax
-        loop.call_later(tmax, self.terminate, proc, task)
+        handle = loop.call_later(tmax, self.terminate, task.id)
         task.state = State.running
         (self.folder / f'local-{task.id}-0').write_text('')  # running
+
         await proc.wait()
-        self.tasks.remove(task)
+        print('END', task.id)
+        handle.cancel()
+
+        del self.tasks[task.id]
+        del self.processes[task.id]
+
         if proc.returncode == 0:
-            for t in self.tasks:
+            for t in self.tasks.values():
                 if task.dname in t.deps:
                     t.deps.remove(task.dname)
             state = 1
@@ -139,17 +195,21 @@ class Server:
                 state = 3
             else:
                 state = 2
-            task.cancel_dependents(self.tasks)
-            self.tasks = [task for task in self.tasks
-                          if task.state != 'CANCELED']
+            task.cancel_dependents(list(self.tasks.values()))
+            self.tasks = {id: task for id, task in self.tasks.items()
+                          if task.state != 'CANCELED'}
+
         (self.folder / f'local-{task.id}-{state}').write_text('')
+
         self.kick()
 
-    def terminate(self, proc: Any, task: Task) -> None:
-        if proc.returncode is None:
+    def terminate(self, id: int, state: State = State.TIMEOUT) -> None:
+        print('Terminate', id)
+        proc = self.processes.get(id)
+        if proc and proc.returncode is None:
             proc.terminate()
-            task.state = State.TIMEOUT
+            self.tasks[id].state = state
 
 
 if __name__ == '__main__':
-    asyncio.run(Server().main())
+    asyncio.run(Server(Configuration.read()).main())